Created Special DAO Interfaces For Spring Beans
+ Because the various methods for DAO implementations are divided up
among four interfaces in the repo package, Spring's proxy mechanism can
cause issues when using the container to inject a DAO class.
+ GenericEntityDaoBean and GenericJoinHolderDaoBean interfaces solve
this problem by providing single, unified interfaces for DAO classes to
implement. This allows for a single, polymorphic type to be declared and
used to make injection from the Spring container and the use of those
injected DAO beans as simple as possible.