Transformed Gender Into Enum Instead of Char
+ JSP EL cannot do comparisons between char variables. Ultimately
decided that if it needed to be changed, it should be changed to an
Enum.
+ Changed database schema to accept larger character strings in gender
column.
+ Created Gender enum with the values MALE, FEMALE, OTHER, and UNKNOWN.
+ Changed Student entity's gender field data type to the new Gender
enum.
+ Mapped the gender field using the @Enumeration annotation to indicate
it should be persisted as the String value of the enum.
+ Updated AdminStudentController to add the values of the Gender enum to
the model for the student form page, so that they can be displayed in
the <select> box.
+ Updated student-form.jsp to retrieve the Gender enum's values for the
<select> box, and set the appropriate one as selected based on the
underlying Student entity.
+ Updated createStudent convenience method in SchoolDataService to
accept the new data type for the gender field.
+ Updated StudentTest, JoinHolderDaoIT, StudentDaoIT, and
SchoolDataServiceIT to match the new gender data type.