Changed HashCode and Equals Methods in Address Entities
+ Current comparisons in Address classes were based on IDs. Since
Address entities don't get IDs until they're persisted, and before that
happens they get added to a HashSet, this breaks the set because all
entities are viewed as being equal.
+ Changed hashCode() and equals() to abstract methods in Address
abstract class.
+ Implemented hashCode() in Address subclasses to be the hashCode value
of a concatanated String of all the values of all the fields of the
class, with an additional String for the type (Student/Admin) of the
class.
+ Implemented equals() in Address subclasses to be a comparison of the
class type and of all the fields of the class.